Cleaned some files and added text files from the scenarios.
Removed .NET warnings
modified   ctp2_code/gfx/tilesys/tiledmap.cpp
modified   ctp2_code/gfx/tilesys/tiledmap.h
modified   ctp2_code/gfx/tilesys/tiledraw.cpp
modified   ctp2_code/gfx/tilesys/TileInfo.h

Cleaned Peter's turns to next pop design and improved the running time of the TurnsToNextPop method.
modified   ctp2_code/gs/gameobj/CityData.cpp
modified   ctp2_code/gs/gameobj/citydata.h

Tried to fix the turns to next pop on map feature, unfortunatly it doesn't work very well, for instance if you modify the number of specialists, you have to close and open the city manager again. And when I load an old savegame the number of turns is wrong, for some reason they don't consider specialists.
modified   ctp2_code/ui/interface/citywindow.cpp

Added text files from scenarios. Espeacilly the const.txt's, it was a bad idea to remove them, because now we can't track the changes and a future playtest wouldn't restore these files to their original status. And obviously this is needed to be able to play the Alex scenario.
added      Scenarios/AlexanderTheGreat/scen0000/default/aidata/UnitBuildLists.txt
added      Scenarios/AlexanderTheGreat/scen0000/default/gamedata/AG_func.slc
added      Scenarios/AlexanderTheGreat/scen0000/default/gamedata/AG_mod.slc
added      Scenarios/AlexanderTheGreat/scen0000/default/gamedata/buildings.txt
added      Scenarios/AlexanderTheGreat/scen0000/default/gamedata/citysize0.txt
added      Scenarios/AlexanderTheGreat/scen0000/default/gamedata/citysize1.txt
added      Scenarios/AlexanderTheGreat/scen0000/default/gamedata/citysize2.txt
added      Scenarios/AlexanderTheGreat/scen0000/default/gamedata/citysize3.txt
added      Scenarios/AlexanderTheGreat/scen0000/default/gamedata/citysize4.txt
added      Scenarios/AlexanderTheGreat/scen0000/default/gamedata/citysize5.txt
added      Scenarios/AlexanderTheGreat/scen0000/default/gamedata/civilisation.txt
added      Scenarios/AlexanderTheGreat/scen0000/default/gamedata/concept.txt
added      Scenarios/AlexanderTheGreat/scen0000/default/gamedata/Const.txt
added      Scenarios/AlexanderTheGreat/scen0000/default/gamedata/DiffDB.txt
added      Scenarios/AlexanderTheGreat/scen0000/default/gamedata/govern.txt
added      Scenarios/AlexanderTheGreat/scen0000/default/gamedata/newsprite.txt
added      Scenarios/AlexanderTheGreat/scen0000/default/gamedata/terrain.txt
added      Scenarios/AlexanderTheGreat/scen0000/default/gamedata/tileimp.txt
added      Scenarios/AlexanderTheGreat/scen0000/default/gamedata/Wonder.txt
added      Scenarios/AlexanderTheGreat/scen0000/german
added      Scenarios/AlexanderTheGreat/scen0000/german/gamedata
added      Scenarios/AlexanderTheGreat/scen0000/german/gamedata/civ_str.txt
added      Scenarios/AlexanderTheGreat/scen0000/german/gamedata/gl_str.txt
added      Scenarios/AlexanderTheGreat/scen0000/german/gamedata/Great_Library.txt
added      Scenarios/AlexanderTheGreat/scen0000/german/gamedata/scen_str.txt
added      Scenarios/AlexanderTheGreat/scen0000/german/gamedata/str_loc.txt
added      Scenarios/AlexanderTheGreat/scen0000/scenicon.tga
added      Scenarios/MagnificentSamurai/scen0000/default/aidata/diplomacy.txt
added      Scenarios/MagnificentSamurai/scen0000/default/aidata/Goals.txt
added      Scenarios/MagnificentSamurai/scen0000/default/aidata/strategies.txt
added      Scenarios/MagnificentSamurai/scen0000/default/gamedata/civilisation.txt
added      Scenarios/MagnificentSamurai/scen0000/default/gamedata/concept.txt
added      Scenarios/MagnificentSamurai/scen0000/default/gamedata/Const.txt
added      Scenarios/MagnificentSamurai/scen0000/default/gamedata/goodsID.txt
added      Scenarios/MagnificentSamurai/scen0000/default/gamedata/newsprite.txt
added      Scenarios/MagnificentSamurai/scen0000/default/gamedata/SS_msg.slc
added      Scenarios/MagnificentSamurai/scen0000/default/gamedata/uniticon.txt
added      Scenarios/MagnificentSamurai/scen0000/default/gamedata/Units.txt
added      Scenarios/MagnificentSamurai/scen0000/german
added      Scenarios/MagnificentSamurai/scen0000/german/gamedata
added      Scenarios/MagnificentSamurai/scen0000/german/gamedata/civ_str.txt
added      Scenarios/MagnificentSamurai/scen0000/german/gamedata/gl_str.txt
added      Scenarios/MagnificentSamurai/scen0000/german/gamedata/Great_Library.txt
added      Scenarios/MagnificentSamurai/scen0000/german/gamedata/scen_str.txt
added      Scenarios/MagnificentSamurai/scen0000/german/gamedata/str_loc.txt
added      Scenarios/NuclearDetente/scen0000/default/aidata
added      Scenarios/NuclearDetente/scen0000/default/aidata/diplomacy.txt
added      Scenarios/NuclearDetente/scen0000/default/aidata/strategies.txt
added      Scenarios/NuclearDetente/scen0000/default/gamedata/civilisation.txt
added      Scenarios/NuclearDetente/scen0000/default/gamedata/concept.txt
added      Scenarios/NuclearDetente/scen0000/default/gamedata/ND_mod.slc
added      Scenarios/NuclearDetente/scen0000/default/gamedata/ND_msg.slc
added      Scenarios/NuclearDetente/scen0000/default/gamedata/uniticon.txt
added      Scenarios/NuclearDetente/scen0000/default/gamedata/Units.txt
added      Scenarios/NuclearDetente/scen0000/german
added      Scenarios/NuclearDetente/scen0000/german/gamedata
added      Scenarios/NuclearDetente/scen0000/german/gamedata/civ_str.txt
added      Scenarios/NuclearDetente/scen0000/german/gamedata/gl_str.txt
added      Scenarios/NuclearDetente/scen0000/german/gamedata/Great_Library.txt
added      Scenarios/NuclearDetente/scen0000/german/gamedata/scen_str.txt
added      Scenarios/NuclearDetente/scen0000/german/gamedata/str_loc.txt

git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@340 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/gfx/tilesys/tiledraw.cpp
@@ -47,6 +47,7 @@
 //   those of the civs you have no contact to. - Mar. 4th 2005 Martin Gühmann
 // - Added option to draw wonders on top of roads.
 // - PFT 29 mar 05, show # turns until city next grows a pop.
+// - Removed .NET warnings. - April 23rd 2005 Martin Gühmann
 //
 //----------------------------------------------------------------------------
 
@@ -428,7 +429,7 @@ bool TiledMap::DrawImprovementsLayer(aui_Surface *surface, MapPoint &pos, sint32
 
 				if(rec!=NULL && effect!=NULL){
 					if(!rec->GetClassRoad() && !rec->GetClassOceanRoad()){
-						data = m_tileSet->GetImprovementData(effect->GetTilesetIndex());
+						data = m_tileSet->GetImprovementData(static_cast<uint16>(effect->GetTilesetIndex()));
 						DrawAnImprovement(surface,data,x,y,fog);
 					}
 				}
@@ -1435,7 +1436,7 @@ if (ypos >= surface->Height() - k_TILE_PIXEL_HEIGHT) return 0;
 	t2 = m_tileSet->GetTransitionData(tilesetIndex_short, tileInfo->GetTransition(2), 2);
 	t3 = m_tileSet->GetTransitionData(tilesetIndex_short, tileInfo->GetTransition(3), 3);
 	
-	transitionBuffer = m_tileSet->GetBaseTile((tilesetIndex * 100) + 99);
+	transitionBuffer = m_tileSet->GetBaseTile(static_cast<uint16>((tilesetIndex * 100) + 99));
 	if(transitionBuffer) {
 		transData = transitionBuffer->GetTileData();
 		transDataPtr = transData;
@@ -1646,7 +1647,7 @@ void TiledMap::DrawBlendedTileScaled(aui_Surface *surface, MapPoint &pos, sint32
 	t2 = m_tileSet->GetTransitionData(tilesetIndex_short, tileInfo->GetTransition(2), 2);
 	t3 = m_tileSet->GetTransitionData(tilesetIndex_short, tileInfo->GetTransition(3), 3);
 
-	transitionBuffer = m_tileSet->GetBaseTile((tilesetIndex * 100) + 99);
+	transitionBuffer = m_tileSet->GetBaseTile(static_cast<uint16>((tilesetIndex * 100) + 99));
 	if(transitionBuffer) {
 		transData = transitionBuffer->GetTileData();
 		transDataPtr = transData;
@@ -4143,7 +4144,7 @@ void TiledMap::DrawTransitionTile(aui_Surface *surface, MapPoint &pos, sint32 xp
 	tileData[2] = m_tileSet->GetTransitionData(tilesetIndex_short, tileInfo->GetTransition(2), 2);
 	tileData[3] = m_tileSet->GetTransitionData(tilesetIndex_short, tileInfo->GetTransition(3), 3);
 	
-	transitionBuffer = m_tileSet->GetBaseTile((tilesetIndex * 100) + 99);
+	transitionBuffer = m_tileSet->GetBaseTile(static_cast<uint16>((tilesetIndex * 100) + 99));
 	if(transitionBuffer) {
 		transData = transitionBuffer->GetTileData();
 		transDataPtr = transData;
@@ -4378,7 +4379,7 @@ void TiledMap::DrawTransitionTileScaled(aui_Surface *surface, MapPoint &pos, sin
 	t2 = m_tileSet->GetTransitionData(tilesetIndex_short, tileInfo->GetTransition(2), 2);
 	t3 = m_tileSet->GetTransitionData(tilesetIndex_short, tileInfo->GetTransition(3), 3);
 	
-	transitionBuffer = m_tileSet->GetBaseTile((tilesetIndex * 100) + 99);
+	transitionBuffer = m_tileSet->GetBaseTile(static_cast<uint16>((tilesetIndex * 100) + 99));
 	if(transitionBuffer) {
 		transData = transitionBuffer->GetTileData();
 		transDataPtr = transData;
